move  test cell serializer into test_cell
